+++ forui/lib/src/widgets/text_field/text_form_field.dart
@@ -6,8 +6,7 @@ class _Field extends FormField<String> {
     FTextField parent,
     FTextFieldStyle style,
     FTextFieldStateStyle stateStyle,
-  ) =>
-      InputDecoration(
+  ) => InputDecoration(
         suffixIcon: parent.suffix,
         // See https://stackoverflow.com/questions/70771410/flutter-how-can-i-remove-the-content-padding-for-error-in-textformfield
         prefix: Padding(padding: EdgeInsets.only(left: style.contentPadding.left)),
@@ -18,13 +17,6 @@ class _Field extends FormField<String> {
             ? null
             : DefaultTextStyle.merge(style: stateStyle.descriptionTextStyle, child: parent.description!),
         helperStyle: stateStyle.descriptionTextStyle,
-        error: state.errorText == null
-            ? null
-            : DefaultTextStyle.merge(
-                style: stateStyle.descriptionTextStyle,
-                child: parent.errorBuilder(state.context, state.errorText!),
-              ),
-        errorStyle: stateStyle.descriptionTextStyle,
         disabledBorder: OutlineInputBorder(
           borderSide: BorderSide(
             color: style.disabledStyle.unfocusedStyle.color,
@@ -147,6 +139,19 @@ class _Field extends FormField<String> {
                     canRequestFocus: parent.canRequestFocus,
                     spellCheckConfiguration: parent.spellCheckConfiguration,
                     magnifierConfiguration: parent.magnifierConfiguration,
+                  ),
+                  AnimatedSwitcher(
+                    duration: style.errorStyle.animationDuration,
+                    child: switch (state.errorText) {
+                      null => const SizedBox(),
+                      final error => Padding(
+                        padding: const EdgeInsets.only(top: 7, bottom: 4),
+                        child: DefaultTextStyle.merge(
+                          style: style.errorStyle.errorTextStyle,
+                          child: Text(error),
+                        ),
+                      ),
+                    },
                   ),
                 ],
               ),